+++ src/dswx_sar/refine_with_bimodality.py
@@ -21,7 +21,7 @@ from dswx_sar.dswx_runconfig import (_get_parser,
                                      RunConfig,
                                      DSWX_S1_POL_DICT)
 
-from dswx_sar.dswx_sar_helper import jacobian_bimodal, jacobian_trimodal
+from dswx_sar.multimodal_jacobian import jacobian_bimodal
 
 logger = logging.getLogger('dswx_sar')
